Automatic Generation of gMCP Reports

Creates a LaTeX file with a gMCP Report.

gMCPReport(object, file = "", ...)

Arguments

object

A graph of class graphMCP or an object of class gMCPResult.

file

A connection, or a character string naming the file to print to. If "" (the default), the report is printed to the standard output connection (the console unless redirected by sink). If it is "|cmd", the output is piped to the command given by cmd, by opening a pipe connection [taken from the manual page of cat, which is called in this function].

...

Arguments to be passed to method graph2latex like package and scale.

Value

None (invisible NULL).

Details

This function uses cat and graph2latex.

References

The TikZ and PGF Packages Manual for version 2.00, Till Tantau, https://www.ctan.org/pkg/pgf/

See also

Author

Kornelius Rohmeyer rohmeyer@small-projects.de

Examples

g <- BretzEtAl2011()

result <- gMCP(g, pvalues=c(0.1, 0.008, 0.005, 0.15, 0.04, 0.006))

gMCPReport(result)
#> \documentclass[11pt]{article}
#> \usepackage{tikz}
#> \usetikzlibrary{decorations,arrows,shapes}
#> \begin{document}
#> \subsection*{Initial graph}
#> \begin{tikzpicture}
#> 
#> \node (H11) at (125bp,-125bp)[draw,circle split,fill=green!80] {$H11$ \nodepart{lower} $\frac{1}{3}$};
#> \node (H21) at (325bp,-125bp)[draw,circle split,double,fill=green!80] {$H21$ \nodepart{lower} $\frac{1}{3}$};
#> \node (H31) at (525bp,-125bp)[draw,circle split,double,fill=green!80] {$H31$ \nodepart{lower} $\frac{1}{3}$};
#> \node (H12) at (125bp,-325bp)[draw,circle split,fill=green!80] {$H12$ \nodepart{lower} $0$};
#> \node (H22) at (325bp,-325bp)[draw,circle split,fill=green!80] {$H22$ \nodepart{lower} $0$};
#> \node (H32) at (525bp,-325bp)[draw,circle split,fill=green!80] {$H32$ \nodepart{lower} $0$};
#> \draw [->,line width=1pt] (H11.17) arc(107:90:260bp) node[fill=blue!20] {$\frac{1}{2}$} arc(90:75:260bp) to (H21);
#> \draw [->,line width=1pt] (H11) to[auto] node[near start,above,fill=blue!20] {$\frac{1}{2}$} (H12);
#> \draw [->,line width=1pt] (H21.197) arc(287:270:260bp) node[fill=blue!20] {$\frac{1}{3}$} arc(270:255:260bp) to (H11);
#> \draw [->,line width=1pt] (H21.343) arc(253:270:260bp) node[fill=blue!20] {$\frac{1}{3}$} arc(270:285:260bp) to (H31);
#> \draw [->,line width=1pt] (H21) to[auto] node[near start,above,fill=blue!20] {$\frac{1}{3}$} (H22);
#> \draw [->,line width=1pt] (H31.163) arc(73:90:260bp) node[fill=blue!20] {$\frac{1}{2}$} arc(90:105:260bp) to (H21);
#> \draw [->,line width=1pt] (H31) to[auto] node[near start,above,fill=blue!20] {$\frac{1}{2}$} (H32);
#> \draw [->,line width=1pt] (H12) to (175bp, -275bp) node[fill=blue!20] {$1$} to (H21);
#> \draw [->,line width=1pt] (H22) to (275bp, -275bp) node[fill=blue!20] {$\frac{1}{2}$} to (H11);
#> \draw [->,line width=1pt] (H22) to (375bp, -275bp) node[fill=blue!20] {$\frac{1}{2}$} to (H31);
#> \draw [->,line width=1pt] (H32) to (475bp, -275bp) node[fill=blue!20] {$1$} to (H21);
#> \end{tikzpicture}
#> 
#> \subsection*{P-Values}
#> \begin{table}[ht]
#> \begin{center}
#> \begin{tabular}{rrrrrr}
#> \hline
#> 
#> H11&H21&H31&H12&H22&H32 \\
#> \hline
#>  0.10000&0.00800&0.00500&0.15000&0.04000&0.00600\\
#> \hline
#>  
#> \end{tabular}
#> \end{center}
#> \end{table}
#> \subsection*{Adjusted p-values}
#> \begin{table}[ht]
#> \begin{center}
#> \begin{tabular}{rrrrrr}
#> \hline
#> 
#> H11&H21&H31&H12&H22&H32 \\
#> \hline
#>  0.12000&0.01600&0.01500&0.15000&0.12000&0.02250\\
#> \hline
#>  
#> \end{tabular}
#> \end{center}
#> \end{table}
#> \subsection*{Rejected Hypotheses with $\alpha=0.05$}
#> \begin{table}[ht]
#> \begin{center}
#> \begin{tabular}{rrrrrr}
#> \hline
#> 
#> H11&H21&H31&H12&H22&H32 \\
#> \hline
#>  FALSE&TRUE&TRUE&FALSE&FALSE&TRUE\\
#> \hline
#>  
#> \end{tabular}
#> \end{center}
#> \end{table}
#> \subsection*{Graph in Step 2}
#> \begin{tikzpicture}
#> 
#> \node (H11) at (125bp,-125bp)[draw,circle split,fill=green!80] {$H11$ \nodepart{lower} $\frac{1}{3}$};
#> \node (H21) at (325bp,-125bp)[draw,circle split,double,fill=green!80] {$H21$ \nodepart{lower} $\frac{1}{2}$};
#> \node (H31) at (525bp,-125bp)[draw,circle split,fill=red!80] {$H31$ \nodepart{lower} $0$};
#> \node (H12) at (125bp,-325bp)[draw,circle split,fill=green!80] {$H12$ \nodepart{lower} $0$};
#> \node (H22) at (325bp,-325bp)[draw,circle split,fill=green!80] {$H22$ \nodepart{lower} $0$};
#> \node (H32) at (525bp,-325bp)[draw,circle split,double,fill=green!80] {$H32$ \nodepart{lower} $\frac{1}{6}$};
#> \draw [->,line width=1pt] (H11.17) arc(107:90:260bp) node[fill=blue!20] {$\frac{1}{2}$} arc(90:75:260bp) to (H21);
#> \draw [->,line width=1pt] (H11) to[auto] node[near start,above,fill=blue!20] {$\frac{1}{2}$} (H12);
#> \draw [->,line width=1pt] (H21.197) arc(287:270:260bp) node[fill=blue!20] {$\frac{2}{5}$} arc(270:255:260bp) to (H11);
#> \draw [->,line width=1pt] (H21) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{2}{5}$} (H22);
#> \draw [->,line width=1pt] (H21) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{1}{5}$} (H32);
#> \draw [->,line width=1pt] (H12) to (175bp, -275bp) node[fill=blue!20] {$1$} to (H21);
#> \draw [->,line width=1pt] (H22) to (275bp, -275bp) node[fill=blue!20] {$\frac{1}{2}$} to (H11);
#> \draw [->,line width=1pt] (H22) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{1}{4}$} (H21);
#> \draw [->,line width=1pt] (H22) to[auto] node[near start,above,fill=blue!20] {$\frac{1}{4}$} (H32);
#> \draw [->,line width=1pt] (H32) to (475bp, -275bp) node[fill=blue!20] {$1$} to (H21);
#> \end{tikzpicture}
#> 
#> \subsection*{Graph in Step 3}
#> \begin{tikzpicture}
#> 
#> \node (H11) at (125bp,-125bp)[draw,circle split,fill=green!80] {$H11$ \nodepart{lower} $\frac{8}{15}$};
#> \node (H21) at (325bp,-125bp)[draw,circle split,fill=red!80] {$H21$ \nodepart{lower} $0$};
#> \node (H31) at (525bp,-125bp)[draw,circle split,fill=red!80] {$H31$ \nodepart{lower} $0$};
#> \node (H12) at (125bp,-325bp)[draw,circle split,fill=green!80] {$H12$ \nodepart{lower} $0$};
#> \node (H22) at (325bp,-325bp)[draw,circle split,fill=green!80] {$H22$ \nodepart{lower} $\frac{1}{5}$};
#> \node (H32) at (525bp,-325bp)[draw,circle split,double,fill=green!80] {$H32$ \nodepart{lower} $\frac{4}{15}$};
#> \draw [->,line width=1pt] (H11) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{5}{8}$} (H12);
#> \draw [->,line width=1pt] (H11) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{1}{4}$} (H22);
#> \draw [->,line width=1pt] (H11) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{1}{8}$} (H32);
#> \draw [->,line width=1pt] (H12) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{2}{5}$} (H11);
#> \draw [->,line width=1pt] (H12) to[auto] node[near start,above,fill=blue!20] {$\frac{2}{5}$} (H22);
#> \draw [->,line width=1pt] (H12) to[auto] node[near start,above,fill=blue!20] {$\frac{1}{5}$} (H32);
#> \draw [->,line width=1pt] (H22) to (275bp, -275bp) node[fill=blue!20] {$\frac{2}{3}$} to (H11);
#> \draw [->,line width=1pt] (H22) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{1}{3}$} (H32);
#> \draw [->,line width=1pt] (H32) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{1}{2}$} (H11);
#> \draw [->,line width=1pt] (H32) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{1}{2}$} (H22);
#> \end{tikzpicture}
#> 
#> \subsection*{Graph in Step 4}
#> \begin{tikzpicture}
#> 
#> \node (H11) at (125bp,-125bp)[draw,circle split,fill=green!80] {$H11$ \nodepart{lower} $\frac{2}{3}$};
#> \node (H21) at (325bp,-125bp)[draw,circle split,fill=red!80] {$H21$ \nodepart{lower} $0$};
#> \node (H31) at (525bp,-125bp)[draw,circle split,fill=red!80] {$H31$ \nodepart{lower} $0$};
#> \node (H12) at (125bp,-325bp)[draw,circle split,fill=green!80] {$H12$ \nodepart{lower} $0$};
#> \node (H22) at (325bp,-325bp)[draw,circle split,fill=green!80] {$H22$ \nodepart{lower} $\frac{1}{3}$};
#> \node (H32) at (525bp,-325bp)[draw,circle split,fill=red!80] {$H32$ \nodepart{lower} $0$};
#> \draw [->,line width=1pt] (H11) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{2}{3}$} (H12);
#> \draw [->,line width=1pt] (H11) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{1}{3}$} (H22);
#> \draw [->,line width=1pt] (H12) to[bend left=15] node[near start,above,fill=blue!20] {$\frac{1}{2}$} (H11);
#> \draw [->,line width=1pt] (H12) to[auto] node[near start,above,fill=blue!20] {$\frac{1}{2}$} (H22);
#> \draw [->,line width=1pt] (H22) to (275bp, -275bp) node[fill=blue!20] {$1$} to (H11);
#> \end{tikzpicture}
#> 
#> \end{document}