Graph2LaTeX

Creates LaTeX code that represents the given graph.

graph2latex(
  graph,
  package = "TikZ",
  scale = 1,
  showAlpha = FALSE,
  alpha = 0.05,
  pvalues,
  fontsize,
  nodeTikZ,
  labelTikZ = "near start,above,fill=blue!20",
  tikzEnv = TRUE,
  offset = c(0, 0),
  fill = list(reject = "red!80", retain = "green!80"),
  fig = FALSE,
  fig.label = NULL,
  fig.caption = NULL,
  fig.caption.short = NULL,
  nodeR = 25,
  scaleText = TRUE
)

Arguments

graph

A graph of class graphMCP.

package

A character string specifying the LaTeX package that should be used. Up to now only TikZ is available.

scale

A numeric scalar specifying a possible scaling of the graph. It is only used if tikzEnv==TRUE. Note that this does only effect the fontsize of the graph if scaleText==FALSE. (Coordinates are interpreted in big points: 72 bp = 1 inch).

showAlpha

Logical whether local alpha levels or weights should be shown.

alpha

An optional numeric argument to specify the type I error rate.

pvalues

If the optional numeric argument pvalues is given, nodes that can be rejected, will be marked.

fontsize

An optional character vector specifying the fontsize for the graph, must be one of "tiny", "scriptsize", "footnotesize", "small", "normalsize", "large", "Large", "LARGE", "huge" or "Huge".

nodeTikZ

A character string with additional arguments for the TikZ node command like for example nodeTikZ="minimum size=2cm".

labelTikZ

A character string with arguments for the TikZ node command within an edge.

tikzEnv

Logical whether the LaTeX code should be wrapped in a TikZ environment.

offset

A numeric of length 2 specifying the x and y offset in the TikZ environment.

fill

A list containing 2 elements reject and retain specifying node fill colour of rejected and retained (or not yet rejected) nodes.

fig

Logical whether a figure environment should be created.

fig.label

Label for figure environment (if fig==TRUE).

fig.caption

Caption for figure environment (if fig==TRUE).

fig.caption.short

Optional short version of fig.caption for list of figures (if fig==TRUE).

nodeR

Radius of nodes (pixel in Java, bp in LaTeX).

scaleText

Only used if scale is unequal 1 and tikzEnv==TRUE. If scaleText is TRUE (the default) a scalebox environment is used. If it is FALSE the optional parameter scale from the tikzpicture environment is used and font size will not change. Note that while you easily can change the scale in the scalebox environment, it is more problematic to adjust the scale in the tikzpicture environment afterwards in the LaTeX document, since for curved edges the parameters are calculated for a certain relative node size which changes if the graph is scaled but the text size stays the same.

Value

A character string that contains LaTeX code representing the given graph.

Details

For details see the given references.

References

The TikZ and PGF Packages Manual for version 2.00, Till Tantau, https://www.ctan.org/pkg/pgf/

See also

Author

Kornelius Rohmeyer rohmeyer@small-projects.de

Examples

g <- BonferroniHolm(5)

graph2latex(g)
#> [1] "\\begin{tikzpicture}\n\n\\node (H1) at (125bp,-225bp)[draw,circle split,fill=green!80] {$H1$ \\nodepart{lower} $\\frac{1}{5}$};\n\\node (H2) at (325bp,-225bp)[draw,circle split,fill=green!80] {$H2$ \\nodepart{lower} $\\frac{1}{5}$};\n\\node (H3) at (525bp,-225bp)[draw,circle split,fill=green!80] {$H3$ \\nodepart{lower} $\\frac{1}{5}$};\n\\node (H4) at (725bp,-225bp)[draw,circle split,fill=green!80] {$H4$ \\nodepart{lower} $\\frac{1}{5}$};\n\\node (H5) at (925bp,-225bp)[draw,circle split,fill=green!80] {$H5$ \\nodepart{lower} $\\frac{1}{5}$};\n\\draw [->,line width=1pt] (H1.42) arc(132:90:125bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:50:125bp) to (H2);\n\\draw [->,line width=1pt] (H1.47) arc(137:90:250bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:45:250bp) to (H3);\n\\draw [->,line width=1pt] (H1.49) arc(139:90:375bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:43:375bp) to (H4);\n\\draw [->,line width=1pt] (H1.50) arc(140:90:500bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:42:500bp) to (H5);\n\\draw [->,line width=1pt] (H2.222) arc(312:270:125bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:230:125bp) to (H1);\n\\draw [->,line width=1pt] (H2.42) arc(132:90:125bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:50:125bp) to (H3);\n\\draw [->,line width=1pt] (H2.47) arc(137:90:250bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:45:250bp) to (H4);\n\\draw [->,line width=1pt] (H2.49) arc(139:90:375bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:43:375bp) to (H5);\n\\draw [->,line width=1pt] (H3.227) arc(317:270:250bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:225:250bp) to (H1);\n\\draw [->,line width=1pt] (H3.222) arc(312:270:125bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:230:125bp) to (H2);\n\\draw [->,line width=1pt] (H3.42) arc(132:90:125bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:50:125bp) to (H4);\n\\draw [->,line width=1pt] (H3.47) arc(137:90:250bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:45:250bp) to (H5);\n\\draw [->,line width=1pt] (H4.229) arc(319:270:375bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:223:375bp) to (H1);\n\\draw [->,line width=1pt] (H4.227) arc(317:270:250bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:225:250bp) to (H2);\n\\draw [->,line width=1pt] (H4.222) arc(312:270:125bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:230:125bp) to (H3);\n\\draw [->,line width=1pt] (H4.42) arc(132:90:125bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(90:50:125bp) to (H5);\n\\draw [->,line width=1pt] (H5.230) arc(320:270:500bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:222:500bp) to (H1);\n\\draw [->,line width=1pt] (H5.229) arc(319:270:375bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:223:375bp) to (H2);\n\\draw [->,line width=1pt] (H5.227) arc(317:270:250bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:225:250bp) to (H3);\n\\draw [->,line width=1pt] (H5.222) arc(312:270:125bp) node[fill=blue!20] {$\\frac{1}{4}$} arc(270:230:125bp) to (H4);\n\\end{tikzpicture}\n"