Function for sample size calculation

Function for sample size calculation

sampSizeCore(
  upperN,
  lowerN = floor(upperN/2),
  targFunc,
  target,
  tol = 0.001,
  alRatio,
  Ntype = c("arm", "total"),
  verbose = FALSE,
  ...
)

Arguments

upperN
targFunc(upperN) should be bigger than target (otherwise upperN is doubled until this is the case).

lowerN
targFunc(lowerN) should be smaller than target (otherwise lowerN is halfed until this is the case).

targFunc
The target (power) function that should be monotonically increasing in n.

target
The target value. The function searches the n with targFunc(n)-target<tol and targFunc(n)>target.

tol
Tolerance: The function searches the n with targFunc(n)-target<tol and targFunc(n)>target.

alRatio
Allocation ratio.

Ntype
Either "arm" or "total".

verbose
Logical, whether verbose output should be printed.

...
...

Value

Integer value n (of type numeric) with targFunc(n)-target<tol and targFunc(n)>target.

Details

For details see the manual and examples.

Author

This function is taken from package DoseFinding under GPL from Bjoern Bornkamp, Jose Pinheiro and Frank Bretz

Examples

f <- function(x){1/100*log(x)}
gMCP:::sampSizeCore(upperN=1000, targFunc=f, target=0.008, verbose=TRUE, alRatio=1)

#> Trying to find a sample size for power 0.008 

#> lower limit for sample size is decreased

#> Upper N: 1000 Upper value 0.0691 
#> Lower N: 2 Lower value 0.0069 
#> 
#> Iter: 1, N = 501, current value = 0.0622
#> Iter: 2, N = 252, current value = 0.0553
#> Iter: 3, N = 127, current value = 0.0484
#> Iter: 4, N = 64, current value = 0.0416
#> Iter: 5, N = 33, current value = 0.035
#> Iter: 6, N = 18, current value = 0.0289
#> Iter: 7, N = 10, current value = 0.023
#> Iter: 8, N = 6, current value = 0.0179
#> Iter: 9, N = 4, current value = 0.0139
#> Iter: 10, N = 3, current value = 0.011

#> $samp.size
#> [1] 3
#> 
#> $target
#> [1] 0.011
#> 
#> attr(,"alRatio")
#> [1] 1
#> attr(,"target")
#> [1] 0.008
#> attr(,"Ntype")
#> [1] "arm"
#> attr(,"class")
#> [1] "sampSize"