Graphical described Multiple Comparison Procedures

This package provides functions and graphical user interfaces for graphical described multiple comparison procedures.

Details

Package:	gMCP

Type:	Package

License:	GPL (>= 2)

The package gMCP helps with the following steps of performing a multiple test procedure:

1. Creating a object of graphMCP that represents a sequentially rejective multiple test procedure. This can be either done directly via the new function or converter functions like matrix2graph at the R command line or by using a graphical user interface started with function graphGUI.

2. Calling gMCP or graphGUI.

3. Exporting the results (optional with all sequential steps) as LaTeX or Word report.

Author

Kornelius Rohmeyer, R code for correlated tests and adaptive designs from Florian Klinglmueller.

Maintainer: Kornelius Rohmeyer rohmeyer@small-projects.de

References

Frank Bretz, Martin Posch, Ekkehard Glimm, Florian Klinglmueller, Willi Maurer, Kornelius Rohmeyer (2011): Graphical approaches for multiple comparison procedures using weighted Bonferroni, Simes or parametric tests. Biometrical Journal 53 (6), pages 894-913, Wiley. http://onlinelibrary.wiley.com/doi/10.1002/bimj.201000239/full

Frank Bretz, Willi Maurer, Werner Brannath, Martin Posch: A graphical approach to sequentially rejective multiple test procedures. Statistics in Medicine 2009 vol. 28 issue 4 page 586-604. http://www.meduniwien.ac.at/fwf_adaptive/papers/bretz_2009_22.pdf

Note

We use the following Java libraries:

Apache Commons Logging under the Apache License, Version 2.0, January 2004, http://commons.apache.org/logging/, Copyright 2001-2007 The Apache Software Foundation

Apache jog4j under Apache License 2.0, http://logging.apache.org/log4j/, Copyright 2007 The Apache Software Foundation

Apache Commons Lang under Apache License 2.0, http://commons.apache.org/lang/, Copyright 2001-2011 The Apache Software Foundation

Apache POI under Apache License 2.0, http://poi.apache.org/, Copyright The Apache Software Foundation

JLaTeXMath under GPL >= 2.0, http://forge.scilab.org/index.php/p/jlatexmath/, Copyright 2004-2007, 2009 Calixte, Coolsaet, Cleemput, Vermeulen and Universiteit Gent

JRI under Lesser General Public License (LGPL) 2.1, http://www.rforge.net/rJava/, Copyright 2004-2007 Simon Urbanek

iText 2.1.4 under LGPL, http://itextpdf.com/, Copyright by Bruno Lowagie

SwingWorker under LGPL, from java.net/projects/swingworker/, Copyright (c) 2005 Sun Microsystems

JXLayer under BSD License, from java.net/projects/jxlayer/, Copyright 2006-2009, Alexander Potochkin

JGoodies Forms under BSD License, http://www.jgoodies.com/freeware/libraries/forms/, Copyright JGoodies Karsten Lentzsch

AFCommons under BSD License, https://web.archive.org/web/20180828002833/http://www.algorithm-forge.com/afcommons/, Copyright (c) 2007-2014 by Kornelius Rohmeyer and Bernd Bischl

JHLIR under BSD License, http://jhlir.r-forge.r-project.org/, Copyright (c) 2008-2014 by Bernd Bischl and Kornelius Rohmeyer

Examples

g5 <- BonferroniHolm(5)
if (FALSE) {
graphGUI("g5")}
gMCP(g5, pvalues=c(0.1,0.2,0.4,0.4,0.4))
#> gMCP-Result
#> 
#> Initial graph:
#> A graphMCP graph
#> H1 (weight=0.2)
#> H2 (weight=0.2)
#> H3 (weight=0.2)
#> H4 (weight=0.2)
#> H5 (weight=0.2)
#> Edges:
#> H1  -( 0.25 )->  H2 
#> H1  -( 0.25 )->  H3 
#> H1  -( 0.25 )->  H4 
#> H1  -( 0.25 )->  H5 
#> H2  -( 0.25 )->  H1 
#> H2  -( 0.25 )->  H3 
#> H2  -( 0.25 )->  H4 
#> H2  -( 0.25 )->  H5 
#> H3  -( 0.25 )->  H1 
#> H3  -( 0.25 )->  H2 
#> H3  -( 0.25 )->  H4 
#> H3  -( 0.25 )->  H5 
#> H4  -( 0.25 )->  H1 
#> H4  -( 0.25 )->  H2 
#> H4  -( 0.25 )->  H3 
#> H4  -( 0.25 )->  H5 
#> H5  -( 0.25 )->  H1 
#> H5  -( 0.25 )->  H2 
#> H5  -( 0.25 )->  H3 
#> H5  -( 0.25 )->  H4 
#> 
#> 
#> P-values:
#>  H1  H2  H3  H4  H5 
#> 0.1 0.2 0.4 0.4 0.4 
#> 
#> Adjusted p-values:
#>  H1  H2  H3  H4  H5 
#> 0.5 0.8 1.0 1.0 1.0 
#> 
#> Alpha: 0.05 
#> 
#> No hypotheses could be rejected.